bugfix: KCM config panel not showing up (#23)
KCM complained when trying to open up the script config panel:

    Plugin does not provide configuration file in expected location

This was due to a behavior change in KWin when loading KCM modules for
plugins. We need to designate the related KCM module manually, using
`X-KDE-ConfigModule` field in metadata

Note that we're using `genericscripted` module, which generates config
dialog using `config.xml` and `config.ui`.